Move all CUDA events into `Buffer`
