+++ cpp/src/buffer/buffer.cpp
@@ -18,21 +18,84 @@ template <typename T>
     RAPIDSMPF_EXPECTS(ptr, "unique pointer cannot be null", std::invalid_argument);
     return ptr;
 }
-
-// Helper to create and record a CUDA event
-cudaEvent_t create_and_record_event(rmm::cuda_stream_view stream) {
-    cudaEvent_t event;
-    RAPIDSMPF_CUDA_TRY(cudaEventCreateWithFlags(&event, cudaEventDisableTiming));
-    RAPIDSMPF_CUDA_TRY(cudaEventRecord(event, stream));
-    return event;
-}
 }  // namespace
+
+/**
+ * @brief CUDA event to provide synchronization among set of chunks.
+ *
+ * This event is used to serve as a synchronization point for a set of chunks
+ * given a user-specified stream.
+ */
+class Event {
+  public:
+    /**
+     * @brief Construct a CUDA event for a given stream.
+     *
+     * @param stream CUDA stream used for device memory operations
+     * @param log Logger to warn if object is destroyed before event is ready.
+     */
+    Event(rmm::cuda_stream_view stream) {
+        RAPIDSMPF_CUDA_TRY(cudaEventCreateWithFlags(&event_, cudaEventDisableTiming));
+        RAPIDSMPF_CUDA_TRY(cudaEventRecord(event_, stream));
+    }
+
+    /**
+     * @brief Destructor for Event.
+     *
+     * Cleans up the CUDA event if one was created. If the event is not done,
+     * it will log a warning.
+     */
+    ~Event() {
+        // Mark as destroying - if we fail, another thread is already destroying
+        bool expected = false;
+        if (!destroying_.compare_exchange_strong(expected, true)) {
+            return;
+        }
+
+        // Finally acquire the mutex and destroy the event
+        std::lock_guard<std::mutex> lock(mutex_);
+        cudaEventDestroy(event_);
+    }
+
+    /**
+     * @brief Check if the CUDA event has been completed.
+     *
+     * @return true if the event has been completed, false otherwise.
+     */
+    [[nodiscard]] bool is_ready() {
+        // Fast path: if done or being destroyed, return immediately
+        if (done_.load(std::memory_order_relaxed)
+            || destroying_.load(std::memory_order_acquire))
+        {
+            return true;
+        }
+
+        // Acquire mutex and check destroying_ again, if being destroyed, return the
+        // previous value of done_.
+        std::lock_guard<std::mutex> lock(mutex_);
+        if (destroying_.load(std::memory_order_acquire)) {
+            return done_.load(std::memory_order_relaxed);
+        }
+
+        // If we're not destroying, check if the event is ready
+        return cudaEventQuery(event_) == cudaSuccess;
+    }
+
+  private:
+    cudaEvent_t event_;  ///< CUDA event used to track device memory allocation
+    // Communicator::Logger&
+    //     log_;  ///< Logger to warn if object is destroyed before event is ready
+    std::atomic<bool> done_{false
+    };  ///< Cache of the event status to avoid unnecessary queries.
+    mutable std::mutex mutex_;  ///< Protects access to event_
+    std::atomic<bool> destroying_{false};  ///< Flag to indicate destruction in progress
+};
 
 Buffer::Buffer(std::unique_ptr<std::vector<uint8_t>> host_buffer, BufferResource* br)
     : br{br},
       size{host_buffer ? host_buffer->size() : 0},
       storage_{std::move(host_buffer)},
-      cuda_event_{nullptr} {
+      event_{nullptr} {
     RAPIDSMPF_EXPECTS(
         std::get<HostStorageT>(storage_) != nullptr, "the host_buffer cannot be NULL"
     );
@@ -47,17 +110,11 @@ Buffer::Buffer(
     : br{br},
       size{device_buffer ? device_buffer->size() : 0},
       storage_{std::move(device_buffer)},
-      cuda_event_{create_and_record_event(stream)} {
+      event_{std::make_unique<Event>(stream)} {
     RAPIDSMPF_EXPECTS(
         std::get<DeviceStorageT>(storage_) != nullptr, "the device buffer cannot be NULL"
     );
     RAPIDSMPF_EXPECTS(br != nullptr, "the BufferResource cannot be NULL");
-}
-
-Buffer::~Buffer() {
-    if (cuda_event_ != nullptr) {
-        cudaEventDestroy(cuda_event_);
-    }
 }
 
 void* Buffer::data() {
@@ -122,7 +179,7 @@ std::unique_ptr<Buffer> Buffer::copy(MemoryType target, rmm::cuda_stream_view st
 
                 // The event is created here instead of the constructor because the
                 // memcpy is async, but the buffer is created on the host.
-                new_buffer->cuda_event_ = create_and_record_event(stream);
+                new_buffer->event_ = std::make_unique<Event>(stream);
 
                 return new_buffer;
             }
@@ -132,18 +189,10 @@ std::unique_ptr<Buffer> Buffer::copy(MemoryType target, rmm::cuda_stream_view st
 }
 
 bool Buffer::is_ready() const {
-    if (cuda_event_ == nullptr) {
+    if (event_ == nullptr) {
         return true;  // No device memory operation was performed
     }
-    cudaError_t status = cudaEventQuery(cuda_event_);
-    if (status == cudaSuccess) {
-        return true;
-    } else if (status == cudaErrorNotReady) {
-        return false;
-    } else {
-        RAPIDSMPF_CUDA_TRY(status);
-        return false;  // This line is unreachable due to the throw above
-    }
+    return event_->is_ready();
 }
 
 }  // namespace rapidsmpf
